requirements.txt
asgiref==3.8.1
certifi==2024.6.2
charset-normalizer==3.3.2
config==0.5.1
decouple==0.0.7
Django==5.0.6
idna==3.7
python-decouple==3.8
python-dotenv==1.0.1
requests==2.32.3
sqlparse==0.5.0
urllib3==2.2.2

